+++ skvideo/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.12.dev0"
+__version__ = "1.1.12"
 
 from .utils import check_output, where
 import os